+++ penny-client/PennyClient/Service/PennyService.swift
@@ -1,119 +1,8 @@
-// swiftlint:disable file_length
 import Foundation
 import Observation
 import SwiftUI
 import UIKit
 import UserNotifications
-
-struct MessagePageCursor: Equatable, Sendable {
-    let createdAt: Date
-    let id: Int
-}
-
-extension MessagePageCursor: CustomStringConvertible {
-    var description: String {
-        "createdAt=\(createdAt.ISO8601Format()), id=\(id)"
-    }
-}
-
-struct MessagePageRequest: Sendable {
-    let limit: Int
-    let before: MessagePageCursor?
-    let filter: MessagePageFilter
-
-    init(limit: Int = 30, before: MessagePageCursor? = nil, filter: MessagePageFilter = .all) {
-        self.limit = limit
-        self.before = before
-        self.filter = filter
-    }
-}
-
-struct MessagePage {
-    let messages: [ChatMessage]
-    let nextCursor: MessagePageCursor?
-    let hasMore: Bool
-}
-
-private struct HistoryPageResult {
-    let payload: MessagesPayload
-    let savedOrUpdatedCount: Int
-}
-
-private struct HistorySyncState: Codable, Equatable {
-    let channelTypes: [String]
-    let includeAttachments: Bool
-    var cursor: String?
-    var requestedCount: Int
-    var savedOrUpdatedCount: Int
-    var remainingCount: Int
-}
-
-enum PennyEmbeddingError: LocalizedError {
-    case unavailable(String)
-    case invalidResponse
-    case disconnected
-
-    var errorDescription: String? {
-        switch self {
-        case .unavailable(let message):
-            return message
-        case .invalidResponse:
-            return "The server returned an invalid embedding."
-        case .disconnected:
-            return "Penny is disconnected."
-        }
-    }
-}
-
-private enum HistorySyncEvent {
-    case page(HistoryPageResult)
-    case error(String)
-}
-
-enum MessagePageFilter: Equatable, Sendable {
-    case all
-    case penny
-    case schedule
-    case chat
-    case notifier
-    case collector
-
-    private static let collectorPrefix = "Collector: "
-
-    var debugDescription: String {
-        switch self {
-        case .all:
-            return "all"
-        case .penny:
-            return "penny"
-        case .schedule:
-            return "schedule"
-        case .chat:
-            return "chat"
-        case .notifier:
-            return "notifier"
-        case .collector:
-            return "collector"
-        }
-    }
-
-    func includes(_ message: ChatMessage) -> Bool {
-        switch self {
-        case .all:
-            return true
-        case .penny:
-            return ["Penny", "Startup", "Test Push"].contains(message.sourceHint)
-        case .schedule:
-            return message.sourceHint == "Schedule"
-        case .chat:
-            return message.isOutgoing || message.sourceHint == "Chat"
-        case .notifier:
-            return message.sourceHint == "Notifier"
-        case .collector:
-            return message.sourceHint?.hasPrefix(Self.collectorPrefix) == true
-        }
-    }
-}
 
 @MainActor
 @Observable
@@ -124,6 +13,7 @@ final class PennyService {
     private var localMessageID = -1
     private let databaseService: DatabaseService
     private let prefs: Prefs
+    private let logger = OSLogService(category: .pennyService)
     private let encoder = JSONEncoder()
     private let decoder = JSONDecoder()
 
@@ -135,6 +25,7 @@ final class PennyService {
     @ObservationIgnored private var historyResponseContinuation: AsyncStream<HistorySyncEvent>.Continuation?
     @ObservationIgnored private var historyPageIsNewest = false
     @ObservationIgnored private var embeddingContinuations: [String: CheckedContinuation<Data, Error>] = [:]
+    @ObservationIgnored private var pendingResponseTimings: [String: [PendingResponseTiming]] = [:]
 
     var messages: [ChatMessage] = []
     var pendingCount = 0
@@ -268,6 +159,7 @@ extension PennyService {
         historySyncTask = nil
         let pendingEmbeddingRequests = Array(embeddingContinuations.values)
         embeddingContinuations.removeAll()
+        pendingResponseTimings.removeAll()
         for continuation in pendingEmbeddingRequests {
             continuation.resume(throwing: PennyEmbeddingError.disconnected)
         }
@@ -281,14 +173,16 @@ extension PennyService {
     }
 
     func requestMessagePage(_ request: MessagePageRequest) async -> MessagePage {
-        debugLogFrame(
+        logger.debug(
             "message page requested " +
-            "(limit=\(request.limit), before=\(request.before?.description ?? "nil"), filter=\(request.filter.debugDescription))"
+            "(limit=\(request.limit), before=\(request.before?.description ?? "nil"), filter=\(request.filter.debugDescription))",
+            privacy: .public
         )
         let page = databaseService.loadMessagePage(request)
-        debugLogFrame(
+        logger.debug(
             "message page returned " +
-            "(count=\(page.messages.count), hasMore=\(page.hasMore), nextCursor=\(page.nextCursor?.description ?? "nil"))"
+            "(count=\(page.messages.count), hasMore=\(page.hasMore), nextCursor=\(page.nextCursor?.description ?? "nil"))",
+            privacy: .public
         )
         return page
     }
@@ -608,7 +502,7 @@ extension PennyService {
 
     private func handleReceiveFailure(_ error: Error) {
         lastError = "WebSocket receive failed: \(error.localizedDescription)"
-        print(lastError ?? error.localizedDescription)
+        logger.error("WebSocket receive failed: \(error.localizedDescription)", privacy: .public)
         disconnect()
     }
 
@@ -623,10 +517,12 @@ extension PennyService {
 
     private func skipUndecodableMessage(_ data: Data) {
         let type = (try? decoder.decode(ServerMessageType.self, from: data))?.type ?? "unknown"
-        print("Skipping unrecognized server message (type: \(type))")
+        logger.warning("Skipping unrecognized server message (type: \(type))", privacy: .public)
     }
 
     private func apply(_ envelope: ServerEnvelope) {
+        logResponseLatency(for: envelope)
+
         switch envelope {
         case .status(let payload):
             isConnected = payload.connected
@@ -902,8 +798,13 @@ extension PennyService {
         Task {
             do {
                 let data = try encoder.encode(outgoingMessage)
-                debugLogFrame("sent frame (\(data.count) bytes)")
-                try await webSocketClient.send(data)
+                let pendingResponse = recordPendingResponse(for: outgoingMessage)
+                do {
+                    try await webSocketClient.send(data)
+                } catch {
+                    removePendingResponse(pendingResponse)
+                    throw error
+                }
             } catch {
                 await MainActor.run {
                     self.lastError = error.localizedDescription
@@ -912,1259 +813,52 @@ extension PennyService {
         }
     }
 
-    /// Logs a short, non-sensitive frame summary in debug builds only. Never logs frame
-    /// contents: outbound frames carry the device secret and APNs token, inbound frames
-    /// carry chat content, and on device these would land in the unified system log.
-    private func debugLogFrame(_ summary: @autoclosure () -> String) {
-        #if DEBUG
-        print("[PennyService] \(summary())")
-        #endif
+    private struct PendingResponseTiming {
+        let id = UUID()
+        let responseKey: String
+        let startDate = Date()
+    }
+
+    private func recordPendingResponse(for outgoingMessage: ClientMessage) -> PendingResponseTiming? {
+        guard let responseKey = outgoingMessage.expectedResponseLogKey else { return nil }
+        let timing = PendingResponseTiming(responseKey: responseKey)
+        pendingResponseTimings[responseKey, default: []].append(timing)
+        return timing
+    }
+
+    private func removePendingResponse(_ timing: PendingResponseTiming?) {
+        guard let timing,
+              var timings = pendingResponseTimings[timing.responseKey] else { return }
+
+        timings.removeAll { $0.id == timing.id }
+        if timings.isEmpty {
+            pendingResponseTimings[timing.responseKey] = nil
+        } else {
+            pendingResponseTimings[timing.responseKey] = timings
+        }
+    }
+
+    private func logResponseLatency(for envelope: ServerEnvelope) {
+        guard let responseKey = envelope.responseLogKey,
+              var timings = pendingResponseTimings[responseKey],
+              !timings.isEmpty else { return }
+
+        let timing = timings.removeFirst()
+        if timings.isEmpty {
+            pendingResponseTimings[responseKey] = nil
+        } else {
+            pendingResponseTimings[responseKey] = timings
+        }
+
+        let elapsedMS = Int(Date().timeIntervalSince(timing.startDate) * 1_000)
+        logger.debug(
+            "websocket response latency (type=\(envelope.logType), elapsed_ms=\(elapsedMS))",
+            privacy: .public
+        )
     }
 
     private func nextLocalMessageID() -> Int {
         defer { localMessageID -= 1 }
         return localMessageID
     }
-}
-
-typealias PennyWebSocketClient = PennyService
-
-private enum ClientMessage: Encodable {
-    case register(RegisterPayload)
-    case message(content: String)
-    case pullMessages(limit: Int)
-    case historyRequest(limit: Int, before: String?, channelTypes: [String]?, includeAttachments: Bool)
-    case ackMessages(ids: [Int])
-    case embeddingRequest(requestID: String, text: String)
-    case heartbeat
-    case configRequest
-    case configUpdate(key: String, value: String)
-    case schedulesRequest
-    case scheduleAdd(command: String)
-    case scheduleUpdate(scheduleID: Int, promptText: String)
-    case scheduleDelete(scheduleID: Int)
-    case promptLogsRequest(agentName: String?, offset: Int?, query: String?, flaggedOnly: Bool?)
-    case memoriesRequest(query: String?)
-    case memoryDetailRequest(name: String, query: String?)
-    case memoryPageRequest(name: String, section: MemorySection, offset: Int, query: String?)
-    case memoryCreate(
-        name: String,
-        description: String,
-        intent: String,
-        inclusion: MemoryInclusion,
-        recall: MemoryRecall,
-        published: Bool?,
-        extractionPrompt: String?,
-        collectorIntervalSeconds: Int?
-    )
-    case memoryUpdate(
-        name: String,
-        description: String?,
-        intent: String?,
-        inclusion: MemoryInclusion?,
-        recall: MemoryRecall?,
-        published: Bool?,
-        extractionPrompt: String?,
-        collectorIntervalSeconds: Int?
-    )
-    case memoryArchive(name: String)
-    case entryCreate(memory: String, key: String, content: String)
-    case entryUpdate(memory: String, key: String, content: String)
-    case entryDelete(memory: String, key: String)
-    case collectionTrigger(name: String)
-    case cursorSet(name: String, logName: String, lastReadAt: String)
-    case cursorClear(name: String, logName: String)
-    case domainUpdate(domain: String, permission: DomainPermission)
-    case domainDelete(domain: String)
-        case permissionDecision(requestID: String, allowed: Bool)
-
-    // swiftlint:disable:next function_body_length
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-
-        switch self {
-        case .register(let payload):
-            try container.encode("register", forKey: .type)
-            try container.encode(payload.deviceID, forKey: .deviceID)
-            try container.encode(payload.label, forKey: .label)
-            try container.encodeIfPresent(payload.pairingToken, forKey: .pairingToken)
-            try container.encodeIfPresent(payload.deviceSecret, forKey: .deviceSecret)
-            try container.encodeIfPresent(payload.apnsToken, forKey: .apnsToken)
-            try container.encode(payload.apnsEnvironment, forKey: .apnsEnvironment)
-            try container.encode(payload.appVersion, forKey: .appVersion)
-        case .message(let content):
-            try container.encode("message", forKey: .type)
-            try container.encode(content, forKey: .content)
-        case .pullMessages(let limit):
-            try container.encode("pull_messages", forKey: .type)
-            try container.encode(limit, forKey: .limit)
-        case .historyRequest(let limit, let before, let channelTypes, let includeAttachments):
-            try container.encode("history_request", forKey: .type)
-            try container.encode(limit, forKey: .limit)
-            try container.encodeIfPresent(before, forKey: .before)
-            try container.encodeIfPresent(channelTypes, forKey: .channelTypes)
-            try container.encode(includeAttachments, forKey: .includeAttachments)
-        case .ackMessages(let ids):
-            try container.encode("ack_messages", forKey: .type)
-            try container.encode(ids, forKey: .ids)
-        case .embeddingRequest(let requestID, let text):
-            try container.encode("embedding_request", forKey: .type)
-            try container.encode(requestID, forKey: .requestID)
-            try container.encode(text, forKey: .text)
-        case .heartbeat:
-            try container.encode("heartbeat", forKey: .type)
-        case .configRequest:
-            try container.encode("config_request", forKey: .type)
-        case .configUpdate(let key, let value):
-            try container.encode("config_update", forKey: .type)
-            try container.encode(key, forKey: .key)
-            try container.encode(value, forKey: .value)
-        case .schedulesRequest:
-            try container.encode("schedules_request", forKey: .type)
-        case .scheduleAdd(let command):
-            try container.encode("schedule_add", forKey: .type)
-            try container.encode(command, forKey: .command)
-        case .scheduleUpdate(let scheduleID, let promptText):
-            try container.encode("schedule_update", forKey: .type)
-            try container.encode(scheduleID, forKey: .scheduleID)
-            try container.encode(promptText, forKey: .promptText)
-        case .scheduleDelete(let scheduleID):
-            try container.encode("schedule_delete", forKey: .type)
-            try container.encode(scheduleID, forKey: .scheduleID)
-        case .promptLogsRequest(let agentName, let offset, let query, let flaggedOnly):
-            try container.encode("prompt_logs_request", forKey: .type)
-            try container.encodeIfPresent(agentName, forKey: .agentName)
-            try container.encodeIfPresent(offset, forKey: .offset)
-            try container.encodeIfPresent(query, forKey: .query)
-            try container.encodeIfPresent(flaggedOnly, forKey: .flaggedOnly)
-        case .memoriesRequest(let query):
-            try container.encode("memories_request", forKey: .type)
-            try container.encodeIfPresent(query, forKey: .query)
-        case .memoryDetailRequest(let name, let query):
-            try container.encode("memory_detail_request", forKey: .type)
-            try container.encode(name, forKey: .name)
-            try container.encodeIfPresent(query, forKey: .query)
-        case .memoryPageRequest(let name, let section, let offset, let query):
-            try container.encode("memory_page_request", forKey: .type)
-            try container.encode(name, forKey: .name)
-            try container.encode(section, forKey: .section)
-            try container.encode(offset, forKey: .offset)
-            try container.encodeIfPresent(query, forKey: .query)
-        case .memoryCreate(
-            let name,
-            let description,
-            let intent,
-            let inclusion,
-            let recall,
-            let published,
-            let extractionPrompt,
-            let collectorIntervalSeconds
-        ):
-            try container.encode("memory_create", forKey: .type)
-            try container.encode(name, forKey: .name)
-            try container.encode(description, forKey: .description)
-            try container.encode(intent, forKey: .intent)
-            try container.encode(inclusion, forKey: .inclusion)
-            try container.encode(recall, forKey: .recall)
-            try container.encodeIfPresent(published, forKey: .published)
-            try container.encodeIfPresent(extractionPrompt, forKey: .extractionPrompt)
-            try container.encodeIfPresent(collectorIntervalSeconds, forKey: .collectorIntervalSeconds)
-        case .memoryUpdate(
-            let name,
-            let description,
-            let intent,
-            let inclusion,
-            let recall,
-            let published,
-            let extractionPrompt,
-            let collectorIntervalSeconds
-        ):
-            try container.encode("memory_update", forKey: .type)
-            try container.encode(name, forKey: .name)
-            try container.encodeIfPresent(description, forKey: .description)
-            try container.encodeIfPresent(intent, forKey: .intent)
-            try container.encodeIfPresent(inclusion, forKey: .inclusion)
-            try container.encodeIfPresent(recall, forKey: .recall)
-            try container.encodeIfPresent(published, forKey: .published)
-            try container.encodeIfPresent(extractionPrompt, forKey: .extractionPrompt)
-            try container.encodeIfPresent(collectorIntervalSeconds, forKey: .collectorIntervalSeconds)
-        case .memoryArchive(let name):
-            try container.encode("memory_archive", forKey: .type)
-            try container.encode(name, forKey: .name)
-        case .entryCreate(let memory, let key, let content):
-            try container.encode("entry_create", forKey: .type)
-            try container.encode(memory, forKey: .memory)
-            try container.encode(key, forKey: .key)
-            try container.encode(content, forKey: .content)
-        case .entryUpdate(let memory, let key, let content):
-            try container.encode("entry_update", forKey: .type)
-            try container.encode(memory, forKey: .memory)
-            try container.encode(key, forKey: .key)
-            try container.encode(content, forKey: .content)
-        case .entryDelete(let memory, let key):
-            try container.encode("entry_delete", forKey: .type)
-            try container.encode(memory, forKey: .memory)
-            try container.encode(key, forKey: .key)
-        case .collectionTrigger(let name):
-            try container.encode("collection_trigger", forKey: .type)
-            try container.encode(name, forKey: .name)
-        case .cursorSet(let name, let logName, let lastReadAt):
-            try container.encode("cursor_set", forKey: .type)
-            try container.encode(name, forKey: .name)
-            try container.encode(logName, forKey: .logName)
-            try container.encode(lastReadAt, forKey: .lastReadAt)
-        case .cursorClear(let name, let logName):
-            try container.encode("cursor_clear", forKey: .type)
-            try container.encode(name, forKey: .name)
-            try container.encode(logName, forKey: .logName)
-        case .domainUpdate(let domain, let permission):
-            try container.encode("domain_update", forKey: .type)
-            try container.encode(domain, forKey: .domain)
-            try container.encode(permission, forKey: .permission)
-        case .domainDelete(let domain):
-            try container.encode("domain_delete", forKey: .type)
-            try container.encode(domain, forKey: .domain)
-        case .permissionDecision(let requestID, let allowed):
-            try container.encode("permission_decision", forKey: .type)
-            try container.encode(requestID, forKey: .requestID)
-            try container.encode(allowed, forKey: .allowed)
-        }
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case type
-        case deviceID = "device_id"
-        case label
-        case pairingToken = "pairing_token"
-        case deviceSecret = "device_secret"
-        case apnsToken = "apns_token"
-        case apnsEnvironment = "apns_environment"
-        case appVersion = "app_version"
-        case content
-        case limit
-        case before
-        case channelTypes = "channel_types"
-        case includeAttachments = "include_attachments"
-        case ids
-        case key
-        case value
-        case command
-        case scheduleID = "schedule_id"
-        case promptText = "prompt_text"
-        case agentName = "agent_name"
-        case offset
-        case query
-        case flaggedOnly = "flagged_only"
-        case name
-        case section
-        case description
-        case intent
-        case inclusion
-        case recall
-        case published
-        case extractionPrompt = "extraction_prompt"
-        case collectorIntervalSeconds = "collector_interval_seconds"
-        case memory
-        case logName = "log_name"
-        case lastReadAt = "last_read_at"
-        case domain
-        case permission
-        case requestID = "request_id"
-        case text
-        case allowed
-    }
-}
-
-private struct RegisterPayload {
-    let deviceID: String
-    let label: String
-    let pairingToken: String?
-    let deviceSecret: String?
-    let apnsToken: String?
-    let apnsEnvironment: String
-    let appVersion: String
-
-    static func current(apnsToken: String?) -> RegisterPayload {
-        RegisterPayload(
-            deviceID: DeviceIdentity.stableDeviceID(),
-            label: UIDevice.current.name,
-            pairingToken: "pairing-token",
-            deviceSecret: DeviceIdentity.deviceSecret(),
-            apnsToken: apnsToken,
-            apnsEnvironment: ApnsEnvironment.current.rawValue,
-            appVersion: Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String ?? "1.0.0"
-        )
-    }
-}
-
-/// APNs environment this build's push token was minted for.
-///
-/// A token is only valid against the matching APNs host, so the server must be
-/// told which one to use. Historically hardcoded to `sandbox`, which silently
-/// broke push on TestFlight/App Store builds (those carry a production token).
-///
-/// Derivation: DEBUG builds always use the development (sandbox) environment.
-/// Release builds read the `aps-environment` entitlement from the embedded
-/// provisioning profile — `development` (a dev or ad-hoc signed build, or a
-/// direct device install) maps to sandbox; `production` — as in a TestFlight or
-/// App Store build, or the absence of an embedded profile — maps to production.
-
-private enum ApnsEnvironment: String {
-    case sandbox
-    case production
-
-    var host: String {
-        switch self {
-        case .sandbox:
-            return "api.sandbox.push.apple.com"
-        case .production:
-            return "api.push.apple.com"
-        }
-    }
-
-    static var current: ApnsEnvironment {
-        #if DEBUG
-        return .sandbox
-        #else
-        return embeddedProfileEnvironment() ?? .production
-        #endif
-    }
-
-    private static func embeddedProfileEnvironment() -> ApnsEnvironment? {
-        guard
-            let url = Bundle.main.url(forResource: "embedded", withExtension: "mobileprovision"),
-            let data = try? Data(contentsOf: url),
-            let entitlements = provisioningEntitlements(from: data),
-            let apsEnvironment = entitlements["aps-environment"] as? String
-        else {
-            return nil
-        }
-        return apsEnvironment == "development" ? .sandbox : .production
-    }
-
-    private static func provisioningEntitlements(from data: Data) -> [String: Any]? {
-        // A .mobileprovision is a CMS (PKCS#7) blob wrapping an XML plist; slice
-        // out the plist between the <plist ...> and </plist> markers and parse it.
-        guard
-            let start = data.range(of: Data("<plist".utf8))?.lowerBound,
-            let end = data.range(of: Data("</plist>".utf8))?.upperBound
-        else {
-            return nil
-        }
-        let plistData = data.subdata(in: start..<end)
-        let profile = try? PropertyListSerialization.propertyList(from: plistData, options: [], format: nil)
-        return (profile as? [String: Any])?["Entitlements"] as? [String: Any]
-    }
-}
-
-private struct ServerMessageType: Decodable {
-    let type: String
-}
-
-private enum ServerEnvelope: Decodable {
-    case status(StatusPayload)
-    case registered(RegisteredPayload)
-    case outboxChanged(OutboxChangedPayload)
-    case messages(MessagesPayload)
-    case messagesAcked(MessagesAckedPayload)
-    case embeddingResponse(EmbeddingResponsePayload)
-    case typing(TypingPayload)
-    case configResponse(ConfigResponsePayload)
-    case schedulesResponse(SchedulesResponsePayload)
-    case promptLogsResponse(PromptLogsResponsePayload)
-    case promptLogUpdate(PromptLogUpdatePayload)
-    case runOutcomeUpdate(RunOutcomeUpdatePayload)
-    case memoriesResponse(MemoriesResponsePayload)
-    case memoryDetailResponse(MemoryDetailResponsePayload)
-    case memoryPageResponse(MemoryPageResponsePayload)
-    case memoryChanged(MemoryChangedPayload)
-    case collectionTriggerResult(CollectionTriggerResult)
-    case domainPermissionsSync(DomainPermissionsSyncPayload)
-    case permissionPrompt(PermissionPrompt)
-    case permissionDismiss(PermissionDismissPayload)
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        let type = try container.decode(String.self, forKey: .type)
-
-        switch type {
-        case "status":
-            self = .status(try StatusPayload(from: decoder))
-        case "registered":
-            self = .registered(try RegisteredPayload(from: decoder))
-        case "outbox_changed":
-            self = .outboxChanged(try OutboxChangedPayload(from: decoder))
-        case "messages":
-            self = .messages(try MessagesPayload(from: decoder))
-        case "messages_acked":
-            self = .messagesAcked(try MessagesAckedPayload(from: decoder))
-        case "embedding_response":
-            self = .embeddingResponse(try EmbeddingResponsePayload(from: decoder))
-        case "typing":
-            self = .typing(try TypingPayload(from: decoder))
-        case "config_response":
-            self = .configResponse(try ConfigResponsePayload(from: decoder))
-        case "schedules_response":
-            self = .schedulesResponse(try SchedulesResponsePayload(from: decoder))
-        case "prompt_logs_response":
-            self = .promptLogsResponse(try PromptLogsResponsePayload(from: decoder))
-        case "prompt_log_update":
-            self = .promptLogUpdate(try PromptLogUpdatePayload(from: decoder))
-        case "run_outcome_update":
-            self = .runOutcomeUpdate(try RunOutcomeUpdatePayload(from: decoder))
-        case "memories_response":
-            self = .memoriesResponse(try MemoriesResponsePayload(from: decoder))
-        case "memory_detail_response":
-            self = .memoryDetailResponse(try MemoryDetailResponsePayload(from: decoder))
-        case "memory_page_response":
-            self = .memoryPageResponse(try MemoryPageResponsePayload(from: decoder))
-        case "memory_changed":
-            self = .memoryChanged(try MemoryChangedPayload(from: decoder))
-        case "collection_trigger_result":
-            self = .collectionTriggerResult(try CollectionTriggerResult(from: decoder))
-        case "domain_permissions_sync":
-            self = .domainPermissionsSync(try DomainPermissionsSyncPayload(from: decoder))
-        case "permission_prompt":
-            self = .permissionPrompt(try PermissionPrompt(from: decoder))
-        case "permission_dismiss":
-            self = .permissionDismiss(try PermissionDismissPayload(from: decoder))
-        default:
-            throw DecodingError.dataCorruptedError(forKey: .type, in: container, debugDescription: "Unknown server message type: \(type)")
-        }
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case type
-    }
-}
-
-private struct StatusPayload: Decodable {
-    let connected: Bool
-    let error: String?
-}
-
-private struct RegisteredPayload: Decodable {
-    let deviceID: String
-    let isDefault: Bool
-    let pendingCount: Int
-
-    private enum CodingKeys: String, CodingKey {
-        case deviceID = "device_id"
-        case isDefault = "is_default"
-        case pendingCount = "pending_count"
-    }
-}
-
-private struct OutboxChangedPayload: Decodable {
-    let pendingCount: Int
-
-    private enum CodingKeys: String, CodingKey {
-        case pendingCount = "pending_count"
-    }
-}
-
-private struct MessagesPayload: Decodable {
-    let messages: [ServerChatMessage]
-    let mode: String
-    let nextCursor: String?
-    let hasMore: Bool
-    let remainingCount: Int
-    let attachmentsIncluded: Bool
-
-    private enum CodingKeys: String, CodingKey {
-        case messages
-        case mode
-        case nextCursor = "next_cursor"
-        case hasMore = "has_more"
-        case remainingCount = "remaining_count"
-        case attachmentsIncluded = "attachments_included"
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        messages = try container.decode([ServerChatMessage].self, forKey: .messages)
-        mode = try container.decodeIfPresent(String.self, forKey: .mode) ?? "outbox"
-        nextCursor = try container.decodeIfPresent(String.self, forKey: .nextCursor)
-        hasMore = try container.decodeIfPresent(Bool.self, forKey: .hasMore) ?? false
-        remainingCount = try container.decodeIfPresent(Int.self, forKey: .remainingCount) ?? 0
-        attachmentsIncluded = try container.decodeIfPresent(Bool.self, forKey: .attachmentsIncluded) ?? true
-    }
-}
-
-private struct MessagesAckedPayload: Decodable {
-    let count: Int
-}
-
-private struct EmbeddingResponsePayload: Decodable {
-    let requestID: String
-    let embedding: String?
-    let error: String?
-
-    private enum CodingKeys: String, CodingKey {
-        case requestID = "request_id"
-        case embedding
-        case error
-    }
-}
-
-private struct TypingPayload: Decodable {
-    let active: Bool
-}
-
-struct RuntimeConfigParam: Decodable, Identifiable {
-    var id: String { key }
-    let key: String
-    let value: String
-    let defaultValue: String
-    let description: String
-    let type: String
-    let group: String
-
-    private enum CodingKeys: String, CodingKey {
-        case key
-        case value
-        case defaultValue = "default"
-        case description
-        case type
-        case group
-    }
-}
-
-private struct ConfigResponsePayload: Decodable {
-    let params: [RuntimeConfigParam]
-}
-
-struct ScheduleItem: Decodable, Identifiable {
-    let id: Int
-    let timingDescription: String
-    let promptText: String
-    let cronExpression: String
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case timingDescription = "timing_description"
-        case promptText = "prompt_text"
-        case cronExpression = "cron_expression"
-    }
-}
-
-private struct SchedulesResponsePayload: Decodable {
-    let schedules: [ScheduleItem]
-    let error: String?
-}
-
-enum RunOutcome: String, Codable {
-    case failed
-    case noWork = "no_work"
-    case worked
-    case incomplete
-    case cancelled
-}
-
-enum RunHealthFlag: String, Codable {
-    case noWorkDone = "no_work_done"
-    case noWrites = "no_writes"
-    case incomplete
-    case toolFailures = "tool_failures"
-    case halfFormedSend = "half_formed_send"
-}
-
-struct RunHealth: Decodable {
-    let bailed: Bool
-    let noWrites: Bool
-    let incomplete: Bool
-    let toolFailures: Int
-    let degenerateSend: Bool
-    let flags: [RunHealthFlag]
-    let regressive: Bool
-
-    static let empty = RunHealth(
-        bailed: false,
-        noWrites: false,
-        incomplete: false,
-        toolFailures: 0,
-        degenerateSend: false,
-        flags: [],
-        regressive: false
-    )
-
-    private enum CodingKeys: String, CodingKey {
-        case bailed
-        case noWrites = "no_writes"
-        case incomplete
-        case toolFailures = "tool_failures"
-        case degenerateSend = "degenerate_send"
-        case flags
-        case regressive
-    }
-}
-
-struct PromptLogRun: Decodable, Identifiable {
-    var id: String { runID }
-    let runID: String
-    let agentName: String
-    var promptCount: Int
-    let startedAt: String
-    var endedAt: String
-    var totalDurationMS: Int
-    var totalInputTokens: Int
-    var totalOutputTokens: Int
-    var runOutcome: RunOutcome?
-    var runReason: String?
-    let runTarget: String?
-    let health: RunHealth
-    let record: String
-    var prompts: [PromptLogEntry]
-
-    init(update: PromptLogUpdateEntry) {
-        runID = update.runID
-        agentName = update.agentName
-        promptCount = 1
-        startedAt = update.timestamp
-        endedAt = update.timestamp
-        totalDurationMS = update.durationMS
-        totalInputTokens = update.inputTokens
-        totalOutputTokens = update.outputTokens
-        runOutcome = nil
-        runReason = nil
-        runTarget = update.runTarget
-        health = .empty
-        record = ""
-        prompts = [PromptLogEntry(update: update)]
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case runID = "run_id"
-        case agentName = "agent_name"
-        case promptCount = "prompt_count"
-        case startedAt = "started_at"
-        case endedAt = "ended_at"
-        case totalDurationMS = "total_duration_ms"
-        case totalInputTokens = "total_input_tokens"
-        case totalOutputTokens = "total_output_tokens"
-        case runOutcome = "run_outcome"
-        case runReason = "run_reason"
-        case runTarget = "run_target"
-        case health
-        case record
-        case prompts
-    }
-}
-
-struct PromptLogEntry: Decodable, Identifiable {
-    let id: Int
-    let timestamp: String
-    let model: String
-    let agentName: String
-    let promptType: String
-    let durationMS: Int
-    let inputTokens: Int
-    let outputTokens: Int
-    let runTarget: String?
-    let messages: [JSONValue]
-    let response: JSONValue
-    let thinking: String
-    let hasTools: Bool
-
-    init(update: PromptLogUpdateEntry) {
-        id = update.id
-        timestamp = update.timestamp
-        model = update.model
-        agentName = update.agentName
-        promptType = update.promptType
-        durationMS = update.durationMS
-        inputTokens = update.inputTokens
-        outputTokens = update.outputTokens
-        runTarget = update.runTarget
-        messages = update.messages
-        response = update.response
-        thinking = update.thinking
-        hasTools = update.hasTools
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case timestamp
-        case model
-        case agentName = "agent_name"
-        case promptType = "prompt_type"
-        case durationMS = "duration_ms"
-        case inputTokens = "input_tokens"
-        case outputTokens = "output_tokens"
-        case runTarget = "run_target"
-        case messages
-        case response
-        case thinking
-        case hasTools = "has_tools"
-    }
-}
-
-struct PromptLogUpdateEntry: Decodable, Identifiable {
-    let id: Int
-    let runID: String
-    let timestamp: String
-    let model: String
-    let agentName: String
-    let promptType: String
-    let durationMS: Int
-    let inputTokens: Int
-    let outputTokens: Int
-    let runTarget: String?
-    let messages: [JSONValue]
-    let response: JSONValue
-    let thinking: String
-    let hasTools: Bool
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case runID = "run_id"
-        case timestamp
-        case model
-        case agentName = "agent_name"
-        case promptType = "prompt_type"
-        case durationMS = "duration_ms"
-        case inputTokens = "input_tokens"
-        case outputTokens = "output_tokens"
-        case runTarget = "run_target"
-        case messages
-        case response
-        case thinking
-        case hasTools = "has_tools"
-    }
-}
-
-private struct PromptLogsResponsePayload: Decodable {
-    let runs: [PromptLogRun]
-    let hasMore: Bool
-
-    private enum CodingKeys: String, CodingKey {
-        case runs
-        case hasMore = "has_more"
-    }
-}
-
-private struct PromptLogUpdatePayload: Decodable {
-    let prompt: PromptLogUpdateEntry
-}
-
-private struct RunOutcomeUpdatePayload: Decodable {
-    let runID: String
-    let outcome: RunOutcome
-    let reason: String
-
-    private enum CodingKeys: String, CodingKey {
-        case runID = "run_id"
-        case outcome
-        case reason
-    }
-}
-
-enum MemoryType: String, Codable {
-    case collection
-    case log
-}
-
-enum MemoryInclusion: String, Codable {
-    case always
-    case relevant
-    case never
-}
-
-enum MemoryRecall: String, Codable {
-    case all
-    case relevant
-    case recent
-}
-
-enum MemorySection: String, Codable {
-    case entries
-    case collectorRuns = "collector_runs"
-}
-
-struct MemoryRecord: Decodable, Identifiable {
-    var id: String { name }
-    let name: String
-    let type: MemoryType
-    let description: String
-    let intent: String?
-    let inclusion: MemoryInclusion
-    let recall: MemoryRecall
-    let published: Bool
-    let archived: Bool
-    let extractionPrompt: String?
-    let collectorIntervalSeconds: Int?
-    let lastCollectedAt: String?
-    let entryCount: Int
-
-    private enum CodingKeys: String, CodingKey {
-        case name
-        case type
-        case description
-        case intent
-        case inclusion
-        case recall
-        case published
-        case archived
-        case extractionPrompt = "extraction_prompt"
-        case collectorIntervalSeconds = "collector_interval_seconds"
-        case lastCollectedAt = "last_collected_at"
-        case entryCount = "entry_count"
-    }
-}
-
-struct MemoryEntryRecord: Decodable, Identifiable {
-    let id: Int
-    let key: String?
-    let content: String
-    let author: String
-    let createdAt: String
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case key
-        case content
-        case author
-        case createdAt = "created_at"
-    }
-}
-
-struct CursorRecord: Decodable, Identifiable {
-    var id: String { logName }
-    let logName: String
-    let lastReadAt: String
-
-    private enum CodingKeys: String, CodingKey {
-        case logName = "log_name"
-        case lastReadAt = "last_read_at"
-    }
-}
-
-struct MemoryDetail {
-    let memory: MemoryRecord
-    var entries: [MemoryEntryRecord]
-    var entriesHasMore: Bool
-    var collectorRuns: [PromptLogRun]
-    var collectorRunsHasMore: Bool
-    var cursors: [CursorRecord]
-
-    fileprivate init(payload: MemoryDetailResponsePayload) {
-        memory = payload.memory
-        entries = payload.entries
-        entriesHasMore = payload.entriesHasMore
-        collectorRuns = payload.collectorRuns
-        collectorRunsHasMore = payload.collectorRunsHasMore
-        cursors = payload.cursors
-    }
-}
-
-struct MemoryPage {
-    let name: String
-    let section: MemorySection
-    let entries: [MemoryEntryRecord]
-    let runs: [PromptLogRun]
-    let hasMore: Bool
-
-    fileprivate init(payload: MemoryPageResponsePayload) {
-        name = payload.name
-        section = payload.section
-        entries = payload.entries
-        runs = payload.runs
-        hasMore = payload.hasMore
-    }
-}
-
-private struct MemoriesResponsePayload: Decodable {
-    let memories: [MemoryRecord]
-}
-
-private struct MemoryDetailResponsePayload: Decodable {
-    let memory: MemoryRecord
-    let entries: [MemoryEntryRecord]
-    let entriesHasMore: Bool
-    let collectorRuns: [PromptLogRun]
-    let collectorRunsHasMore: Bool
-    let cursors: [CursorRecord]
-
-    private enum CodingKeys: String, CodingKey {
-        case memory
-        case entries
-        case entriesHasMore = "entries_has_more"
-        case collectorRuns = "collector_runs"
-        case collectorRunsHasMore = "collector_runs_has_more"
-        case cursors
-    }
-}
-
-private struct MemoryPageResponsePayload: Decodable {
-    let name: String
-    let section: MemorySection
-    let entries: [MemoryEntryRecord]
-    let runs: [PromptLogRun]
-    let hasMore: Bool
-
-    private enum CodingKeys: String, CodingKey {
-        case name
-        case section
-        case entries
-        case runs
-        case hasMore = "has_more"
-    }
-}
-
-private struct MemoryChangedPayload: Decodable {
-    let name: String?
-}
-
-struct CollectionTriggerResult: Decodable {
-    let name: String
-    let success: Bool
-    let message: String
-}
-
-enum DomainPermission: String, Codable {
-    case allowed
-    case blocked
-}
-
-struct DomainPermissionEntry: Decodable, Identifiable {
-    var id: String { domain }
-    let domain: String
-    let permission: DomainPermission
-}
-
-private struct DomainPermissionsSyncPayload: Decodable {
-    let permissions: [DomainPermissionEntry]
-}
-
-struct PermissionPrompt: Decodable, Identifiable {
-    var id: String { requestID }
-    let requestID: String
-    let domain: String
-    let url: String
-
-    private enum CodingKeys: String, CodingKey {
-        case requestID = "request_id"
-        case domain
-        case url
-    }
-}
-
-private struct PermissionDismissPayload: Decodable {
-    let requestID: String
-
-    private enum CodingKeys: String, CodingKey {
-        case requestID = "request_id"
-    }
-}
-
-enum JSONValue: Codable, Equatable {
-    case string(String)
-    case number(Double)
-    case bool(Bool)
-    case object([String: JSONValue])
-    case array([JSONValue])
-    case null
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        if container.decodeNil() {
-            self = .null
-        } else if let value = try? container.decode(Bool.self) {
-            self = .bool(value)
-        } else if let value = try? container.decode(Double.self) {
-            self = .number(value)
-        } else if let value = try? container.decode(String.self) {
-            self = .string(value)
-        } else if let value = try? container.decode([JSONValue].self) {
-            self = .array(value)
-        } else if let value = try? container.decode([String: JSONValue].self) {
-            self = .object(value)
-        } else {
-            throw DecodingError.dataCorruptedError(in: container, debugDescription: "Unsupported JSON value")
-        }
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        switch self {
-        case .string(let value):
-            try container.encode(value)
-        case .number(let value):
-            try container.encode(value)
-        case .bool(let value):
-            try container.encode(value)
-        case .object(let value):
-            try container.encode(value)
-        case .array(let value):
-            try container.encode(value)
-        case .null:
-            try container.encodeNil()
-        }
-    }
-}
-
-private struct ServerChatMessage: Decodable {
-    let id: Int
-    let createdAt: Date
-    let content: String
-    let attachments: [Attachment]
-    let sourceType: String?
-    let sourceName: String?
-    let sourceHint: String?
-    let pushTitle: String?
-    let pushSummary: String?
-    let messageID: Int?
-    let outboxID: Int?
-    let direction: String?
-    let channelType: String?
-    let deviceLabel: String?
-    let deviceIdentifier: String?
-    let parentID: Int?
-    let embedding: Data?
-
-    var canonicalID: Int { messageID ?? id }
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case createdAt = "created_at"
-        case content
-        case attachments
-        case sourceType = "source_type"
-        case sourceName = "source_name"
-        case sourceHint = "source_hint"
-        case pushTitle = "push_title"
-        case pushSummary = "push_summary"
-        case messageID = "message_id"
-        case outboxID = "outbox_id"
-        case direction
-        case channelType = "channel_type"
-        case deviceLabel = "device_label"
-        case deviceIdentifier = "device_identifier"
-        case parentID = "parent_id"
-        case embedding
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        id = try container.decode(Int.self, forKey: .id)
-        content = try container.decode(String.self, forKey: .content)
-        attachments = try container.decodeIfPresent([Attachment].self, forKey: .attachments) ?? []
-        sourceType = try container.decodeIfPresent(String.self, forKey: .sourceType)
-        sourceName = try container.decodeIfPresent(String.self, forKey: .sourceName)
-        sourceHint = try container.decodeIfPresent(String.self, forKey: .sourceHint)
-        pushTitle = try container.decodeIfPresent(String.self, forKey: .pushTitle)
-        pushSummary = try container.decodeIfPresent(String.self, forKey: .pushSummary)
-        messageID = try container.decodeIfPresent(Int.self, forKey: .messageID)
-        outboxID = try container.decodeIfPresent(Int.self, forKey: .outboxID)
-        direction = try container.decodeIfPresent(String.self, forKey: .direction)
-        channelType = try container.decodeIfPresent(String.self, forKey: .channelType)
-        deviceLabel = try container.decodeIfPresent(String.self, forKey: .deviceLabel)
-        deviceIdentifier = try container.decodeIfPresent(String.self, forKey: .deviceIdentifier)
-        parentID = try container.decodeIfPresent(Int.self, forKey: .parentID)
-        if let encodedEmbedding = try container.decodeIfPresent(String.self, forKey: .embedding) {
-            embedding = Data(base64Encoded: encodedEmbedding)
-        } else {
-            embedding = nil
-        }
-
-        let createdAtString = try container.decode(String.self, forKey: .createdAt)
-        createdAt = DateParser.parse(createdAtString) ?? .now
-    }
-}
-
-private struct Attachment: Decodable {
-    let dataURL: String?
-    let url: String?
-    let name: String?
-    let contentType: String?
-
-    var image: UIImage? {
-        guard let dataURL, let data = DataURLDecoder.decode(dataURL) else { return nil }
-        return UIImage(data: data)
-    }
-
-    init(from decoder: Decoder) throws {
-        if let container = try? decoder.singleValueContainer(), let dataURL = try? container.decode(String.self) {
-            self.dataURL = dataURL
-            url = nil
-            name = nil
-            contentType = nil
-            return
-        }
-
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        dataURL = try container.decodeIfPresent(String.self, forKey: .dataURL)
-        url = try container.decodeIfPresent(String.self, forKey: .url)
-        name = try container.decodeIfPresent(String.self, forKey: .name)
-        contentType = try container.decodeIfPresent(String.self, forKey: .contentType)
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case dataURL = "data_url"
-        case url
-        case name
-        case contentType = "content_type"
-    }
-}
-
-struct ImageAttachment: Identifiable {
-    let id = UUID()
-    let image: UIImage
-}
-
-struct ChatMessage: Identifiable {
-    let id: Int
-    let serverID: Int?
-    let createdAt: Date
-    let content: String
-    let sourceHint: String?
-    let channelType: String?
-    let deviceLabel: String?
-    let deviceIdentifier: String?
-    let parentID: Int?
-    let imageAttachmentDataURLs: [String]
-    let imageAttachments: [ImageAttachment]
-    let isOutgoing: Bool
-    let embedding: Data?
-
-    var displayTime: String {
-        createdAt.formatted(date: .omitted, time: .shortened)
-    }
-
-    init(
-        id: Int,
-        serverID: Int?,
-        createdAt: Date,
-        content: String,
-        sourceHint: String?,
-        channelType: String? = nil,
-        deviceLabel: String? = nil,
-        deviceIdentifier: String? = nil,
-        parentID: Int? = nil,
-        imageAttachmentDataURLs: [String] = [],
-        imageAttachments: [ImageAttachment],
-        isOutgoing: Bool,
-        embedding: Data? = nil
-    ) {
-        self.id = id
-        self.serverID = serverID
-        self.createdAt = createdAt
-        self.content = content
-        self.sourceHint = sourceHint
-        self.channelType = channelType
-        self.deviceLabel = deviceLabel
-        self.deviceIdentifier = deviceIdentifier
-        self.parentID = parentID
-        self.imageAttachmentDataURLs = imageAttachmentDataURLs
-        self.imageAttachments = imageAttachments
-        self.isOutgoing = isOutgoing
-        self.embedding = embedding
-    }
-
-    init(model: MessageModel) {
-        id = model.id
-        serverID = model.serverID
-        createdAt = model.createdAt
-        content = model.content
-        sourceHint = model.sourceHint
-        channelType = model.channelType
-        deviceLabel = model.deviceLabel
-        deviceIdentifier = model.deviceIdentifier
-        parentID = model.parentID
-        imageAttachmentDataURLs = model.imageAttachmentDataURLs
-        embedding = model.embedding
-        imageAttachments = model.imageAttachmentDataURLs.compactMap { dataURL in
-            guard let data = DataURLDecoder.decode(dataURL), let image = UIImage(data: data) else { return nil }
-            return ImageAttachment(image: image)
-        }
-        isOutgoing = model.isOutgoing
-    }
-
-    static func local(id: Int, content: String) -> ChatMessage {
-        ChatMessage(id: id, serverID: nil, createdAt: .now, content: content, sourceHint: nil, imageAttachments: [], isOutgoing: true, embedding: nil)
-    }
-
-    fileprivate static func remote(_ message: ServerChatMessage) -> ChatMessage {
-        let imageAttachmentDataURLs = message.attachments.compactMap(\.dataURL)
-        let imageAttachments = message.attachments.compactMap(\.image).map(ImageAttachment.init(image:))
-        return ChatMessage(
-            id: message.canonicalID,
-            serverID: message.canonicalID,
-            createdAt: message.createdAt,
-            content: message.content,
-            sourceHint: message.sourceHint,
-            channelType: message.channelType,
-            deviceLabel: message.deviceLabel,
-            deviceIdentifier: message.deviceIdentifier,
-            parentID: message.parentID,
-            imageAttachmentDataURLs: imageAttachmentDataURLs,
-            imageAttachments: imageAttachments,
-            isOutgoing: message.direction == "incoming",
-            embedding: message.embedding
-        )
-    }
-}
-
-private enum DataURLDecoder {
-    static func decode(_ value: String) -> Data? {
-        let trimmedValue = value.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard let commaIndex = trimmedValue.firstIndex(of: ",") else { return nil }
-
-        let metadata = trimmedValue[..<commaIndex].lowercased()
-        guard metadata.hasPrefix("data:"), metadata.contains(";base64") else { return nil }
-
-        let base64StartIndex = trimmedValue.index(after: commaIndex)
-        let base64 = trimmedValue[base64StartIndex...]
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-            .replacingOccurrences(of: "\n", with: "")
-            .replacingOccurrences(of: "\r", with: "")
-
-        return Data(base64Encoded: base64)
-    }
-}
-
-private enum DateParser {
-    static func parse(_ value: String) -> Date? {
-        if let date = iso8601WithFractionalSeconds.date(from: value) {
-            return date
-        }
-
-        if let date = iso8601.date(from: value) {
-            return date
-        }
-
-        if let date = localTimestampWithFractionalSeconds.date(from: value) {
-            return date
-        }
-
-        return localTimestamp.date(from: value)
-    }
-
-    private static let iso8601: ISO8601DateFormatter = {
-        let formatter = ISO8601DateFormatter()
-        formatter.formatOptions = [.withInternetDateTime]
-        return formatter
-    }()
-
-    private static let iso8601WithFractionalSeconds: ISO8601DateFormatter = {
-        let formatter = ISO8601DateFormatter()
-        formatter.formatOptions = [.withInternetDateTime, .withFractionalSeconds]
-        return formatter
-    }()
-
-    private static let localTimestamp: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.timeZone = TimeZone(secondsFromGMT: 0)
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss"
-        return formatter
-    }()
-
-    private static let localTimestampWithFractionalSeconds: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.timeZone = TimeZone(secondsFromGMT: 0)
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSSSS"
-        return formatter
-    }()
 }
